Integrate differential gene set calculated by all enrichment score matrixes

Wlicox test is perform to all enrichment score matrixes and gene sets with adjusted p value < 0.05 are used to integrated through RRA.

irGSEA.integrate(
  object = NULL,
  group.by = NULL,
  metadata = NULL,
  col.name = NULL,
  method = c("AUCell", "UCell", "singscore", "ssgsea")
)

Arguments

object
Seurat object including enrichment score matrixes.

group.by
Default ident when it set to NULL. You can specify other columns of metadata.

metadata
A vector add to this Seurat object, and then it is set as the ident of Seurat object to perform differential gene sets analysis.

col.name
A name for metadata.

method
A vector to select enrichment score matrixes. Default all enrichment score matrixes

Value

A list including the differential gene sets calculated by enrichment score matrixes through wlicox test. Gene sets with adjusted p value < 0.05 are regarded as statistically significant. Meanwhile, significant differential gene sets are integrated and the results are saved in a data frame named RRA. Among them, Gene sets with p value < 0.05 are statistically significant and common differential in all gene sets enrichment analysis methods.

Examples

if (FALSE) {
# load PBMC dataset by R package SeuratData
library(Seurat)
library(SeuratData)
# download 3k PBMCs from 10X Genomics
InstallData("pbmc3k")
data("pbmc3k.final")
pbmc3k.final <- SeuratObject::UpdateSeuratObject(pbmc3k.final)

# Seurat object
pbmc3k.final <- irGSEA.score(object = pbmc3k.final, assay = "RNA",
slot = "data", msigdb = T, species = "Homo sapiens",
category = "H", geneid = "symbol",
method = c("AUCell", "UCell", "singscore", "ssgsea"), kcdf = 'Gaussian')

# Integrated analysis
result.dge <- irGSEA.integrate(object = pbmc3k.final,
group.by = "seurat_annotations", metadata = NULL, col.name = NULL,
method = c("AUCell","UCell","singscore","ssgsea"))

result.dge2 <- irGSEA.integrate(object = pbmc3k.final, group.by = NULL,
metadata = pbmc3k.final$seurat_annotations, col.name = "ident",
method = c("AUCell","UCell","singscore","ssgsea"))

}